% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadhtx.R
\name{addRD}
\alias{addRD}
\title{add gene-level rowData derived from transcript level rowRanges}
\usage{
addRD(x)
}
\arguments{
\item{x}{result of loadHtxcomp()}
}
\description{
add gene-level rowData derived from transcript level rowRanges
}
