% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bigrnaFiles}
\alias{bigrnaFiles}
\title{bigrnaFiles: vector of files in bigrna}
\format{vector of 186K character strings}
\usage{
bigrnaFiles
}
\description{
bigrnaFiles: vector of files in bigrna
}
\examples{
head(htxcomp::bigrnaFiles)
}
\keyword{datasets}
