% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMeta.R
\name{getMeta}
\alias{getMeta}
\title{utility for adding metadata to htxcomp extraction, using SRAdbV2}
\usage{
getMeta(study.accession = "ERP005938", entity = "full",
  asDataFrame = TRUE)
}
\arguments{
\item{study.accession}{character(1) accession identifier}

\item{entity}{character(1) level of metadata retrieval, can be "full", "study", "experiment",
see SRAdbV2 documentation}

\item{asDataFrame}{logical(1) if FALSE, return a tibble as produced by SRAdbV2, otherwise
return an S4Vectors DataFrame based on atomic columns of the tibble, 
in which non-atomic columns in the tibble are placed in
the metadata component}
}
\value{
either a tibble or DataFrame depending on value of asDataFrame
}
\description{
utility for adding metadata to htxcomp extraction, using SRAdbV2
}
\examples{
requireNamespace("SRAdbV2")
requireNamespace("S4Vectors")
chk = getMeta()
head(names(chk))
names(S4Vectors::metadata(chk)$non.atomic)
chk[1:3,1:5]
}
