% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htxQuery.R
\name{htxQuery}
\alias{htxQuery}
\title{retrieve 'restfulSE' SummarizedExperiment instance for selected studies in htx compendium}
\usage{
htxQuery(study_accessions, ...)
}
\arguments{
\item{study_accessions}{character vector of study accessions}

\item{\dots}{passed to loadHtxcomp}
}
\value{
SummarizedExperiment instance
}
\description{
retrieve 'restfulSE' SummarizedExperiment instance for selected studies in htx compendium
}
\examples{
htxQuery("ERP011411")
}
