% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getatts.R
\name{sampleAtts}
\alias{sampleAtts}
\title{retrieve a normalized data.frame of sample attributes for SRAdbV2 study tibbles}
\usage{
sampleAtts(studyAcc, returnBad = FALSE)
}
\arguments{
\item{studyAcc}{character(1) accession}

\item{returnBad}{logical(1) returns tibble of experiment metadata for which number 
of fields in sample.attributes is less than the number in the majority}
}
\description{
retrieve a normalized data.frame of sample attributes for SRAdbV2 study tibbles
}
\note{
Sometimes a field is omitted for control experiments, and with manual programming
conformant metadata can be produced for these experiments.  Sometimes there is substantial
diversity among sample.attribute fields recorded within a study.
}
